Added SRV class and SRVIter inner class
The SRV class contains all the logic required to use SRV records for
load balancing. The SRV instance can be iterated over to easily use the
SRV records return from DNS. The order the items are iterated over is
based strictly on RFC 2782 for chosing a server.

There is also a command line interface that shell scripts can use for
load balancing.